+++ src/index.routes.js
@@ -7,6 +7,7 @@ const { Router } = require('express');
 const passport = require('passport');
 
 const AuthenticationValidation = require('@middlewares/validations/AuthenticationValidation');
+const CreateUserValidation = require('@middlewares/validations/CreateUserValidation');
 const SchemaPassport = require('@middlewares/Auth');
 
 const routes = new Router();
@@ -16,7 +17,7 @@ routes.post(
   AuthenticationValidation,
   UserAuthenticationController.store,
 );
-userRouter.post('/user', UserController.store);
+userRouter.post('/user', CreateUserValidation, UserController.store);
 
 passport.use(SchemaPassport);
 routes.use(passport.authenticate('jwt', { session: false }));
