+++ src/routes/user.routes.js
@@ -1,10 +1,12 @@
-const UserController = require('@controllers/UserController');
 const { Router } = require('express');
+const UserController = require('@controllers/UserController');
+
+const UpdateUserValidation = require('@middlewares/validations/UpdateUserValidation');
 
 const userRouter = Router();
 
 userRouter.get('/:email', UserController.show);
 userRouter.get('/', UserController.index);
-userRouter.put('/', UserController.update);
+userRouter.put('/', UpdateUserValidation, UserController.update);
 
 module.exports = userRouter;
